requirements.txt
flask
flask-cors
numpy
mysql-connector-python
tensorflow
scikit-learn
pillow